feat: add roles editor components with permissions configuration and roles form
